compilation now working (tests planned)
I plans to test, and finish development in 20-21 december (2017), or in january (2018)

+++ source/gltfViewer.cpp
@@ -219,7 +219,7 @@ namespace SatelliteExample {
 
             // load to unified buffer space
             intptr_t offset = vtbSpace->copyHostBuffer(gltfModel.buffers[i].data, gltfModel.buffers[i].data.size());
-            vtbSpace->addRegionDesc(rt::BufferRegion{ offset, gltfModel.buffers[i].data.size() });
+            vtbSpace->addRegionDesc(rt::BufferRegion{ uint32_t(offset), uint32_t(gltfModel.buffers[i].data.size()) });
         }
 
         // make buffer views
